Added metropolis_sampler

+++ test/inference/test_norm_post_sim.cpp
@@ -17,16 +17,15 @@ TEST_CASE("norm_post_sim check sizes", "[norm_post_sim]") {
 }
 
 TEST_CASE("norm_post_sim check normal distribution", "[norm_post_sim]") {
-    short int size = 3;
+    size_t size           = 3;
     Eigen::VectorXd modes = Eigen::VectorXd::Zero(size);
     Eigen::MatrixXd cov_matrix{Eigen::MatrixXd::Identity(size, size)};
     NormPostSimData data = norm_post_sim(modes, cov_matrix);
 
-    for (int i = 0; i < size; i ++){
-        REQUIRE( (data.mean_est[i] > -0.1) & (data.mean_est[i] < 0.1));
-        REQUIRE( (data.median_est[i] > -0.2) & (data.median_est[i] < 0.2));
-        REQUIRE( (data.upper_95_est[i] > 1));
-        REQUIRE( (data.lower_5_est[i] < -1));
+    for (size_t i = 0; i < size; i++) {
+        REQUIRE((data.mean_est[i] > -0.1) & (data.mean_est[i] < 0.1));
+        REQUIRE((data.median_est[i] > -0.2) & (data.median_est[i] < 0.2));
+        REQUIRE((data.upper_95_est[i] > 1));
+        REQUIRE((data.lower_5_est[i] < -1));
     }
-
 }
